sna: Expand multiplies of two 16-bit values to a full 32-bit range

+++ src/sna/sna_render_inline.h
@@ -90,7 +90,7 @@ is_dirty_gpu(DrawablePtr drawable)
 static inline Bool
 too_small(DrawablePtr drawable)
 {
-	return (drawable->width * drawable->height * drawable->bitsPerPixel <= 8*4096) &&
+	return ((uint32_t)drawable->width * drawable->height * drawable->bitsPerPixel <= 8*4096) &&
 		!is_dirty_gpu(drawable);
 }
